requirements.txt
torch>=2.1.0
transformers>=4.44.0
keras>=3.0.0
keras-hub>=0.15.0
jax[cuda12]>=0.4.30
opencv-python>=4.8.0
faiss-cpu>=1.7.4
gradio>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
nbformat>=5.9.0
